QuestionStoreModule: when initialParamData is provided but incomplete, fall back to default values
extractRealParamValues wasn't working correctly for filter params.

It couldn't make a desired initial state of filters, `"{"filters":[]}"` when providing param.biom_datasets=17298 in the URL:
```
makeDefaultParamValues(question.parameters)
{biom_datasets: "17321", samples_filter_metadata_user: "{"filters":[]}"}
extractRealParamValues(question.parameters, initialParamData)
{biom_datasets: "17298", samples_filter_metadata_user: ""}
```

"" is not a valid state of filters, because it's not valid JSON. There
was an error in the call to JSON.parse in FilterParamUtils.getFilters.