FilterParamNew: add paging to large membership fields
Main rationale of this feature is to make improperly large membership fields usable.
This came up for user datasets in MicrobiomeDB and an example input from
Earth Microbiome Project, containing thousands of unannotated samples.
The UI broke down on this case because rendering of all rows took too
long.

Only page if there are over 100 values: there shouldn't usually be this many.

Offer page sizes of 50, 100, 200, and 1000. Changing pages tries to preserve values that are being shown.

When a search term is entered:
- set lower opacity for a grey out effect
- set pointer events to none
- use page size as maximum search count

Add an "indeterminate" state of a Select All checkbox, when some elements
are selected. Clicking "Select All" on paged rows selects the current
page (similar to behaviour with a search term).
One consequence is that it's not possible to unclick a page. I tried to
make it work but it was difficult to get something non-broken, so went
for consistency in the end.

Fix a small bug in MesaController.jsx, renderPaginationMenu:
the parameter `rowsPerPageOptions` wasn't passed along.

MembershipField.jsx small refactor: make `getRows()` a method similarly
named to `getFilteredRows()`.